+++ test/temml.js
@@ -1235,6 +1235,8 @@ var temml = (function () {
   defineSymbol(math, textord, "$", "\\$");
   defineSymbol(text, textord, "$", "\\$");
   defineSymbol(text, textord, "$", "\\textdollar");
+  defineSymbol(math, textord, "¢", "\\cent");
+  defineSymbol(text, textord, "¢", "\\cent");
   defineSymbol(math, textord, "%", "\\%");
   defineSymbol(text, textord, "%", "\\%");
   defineSymbol(math, textord, "_", "\\_");
@@ -3149,13 +3151,12 @@ var temml = (function () {
     names: ["\\\\"],
     props: {
       numArgs: 0,
-      numOptionalArgs: 1,
-      argTypes: ["size"],
+      numOptionalArgs: 0,
       allowedInText: true
     },
 
     handler({ parser }, args, optArgs) {
-      const size = optArgs[0];
+      const size = parser.gullet.future().text === "[" ? parser.parseSizeGroup(true) : null;
       const newLine = !parser.settings.displayMode;
       return {
         type: "cr",
@@ -4915,7 +4916,6 @@ var temml = (function () {
       "\\mathscr",
       "\\mathsf",
       "\\mathtt",
-      "\\oldstylenums",
 
       // aliases
       "\\Bbb",
@@ -7329,8 +7329,7 @@ var temml = (function () {
     mathfrak: "fraktur",
     mathscr: "script",
     mathsf: "sans-serif",
-    mathtt: "monospace",
-    oldstylenums: "oldstylenums"
+    mathtt: "monospace"
   };
 
   /**
@@ -7400,8 +7399,6 @@ var temml = (function () {
         return "sans-serif"
       case "mathtt":
         return "monospace"
-      case "oldstylenums":
-        return "oldstylenums"
     }
 
     let text = group.text;
@@ -7697,10 +7694,7 @@ var temml = (function () {
       let node;
       if (numberRegEx$1.test(group.text)) {
         const tag = group.mode === "text" ? "mtext" : "mn";
-        if (variant === "oldstylenums") {
-          const ms = new mathMLTree.MathNode("mstyle", [text], ["oldstylenums"]);
-          node = new mathMLTree.MathNode(tag, [ms]);
-        } else if (variant === "italic" || variant === "bold-italic") {
+        if (variant === "italic" || variant === "bold-italic") {
           return italicNumber(text, variant, tag)
         } else {
           if (variant !== "normal") {
@@ -9531,6 +9525,36 @@ var temml = (function () {
     '\u1DBF': 'θ'
   });
 
+  // Used for Unicode input of calligraphic and script letters
+  const asciiFromScript = Object.freeze({
+    "\ud835\udc9c": "A",
+    "\u212c": "B",
+    "\ud835\udc9e": "C",
+    "\ud835\udc9f": "D",
+    "\u2130": "E",
+    "\u2131": "F",
+    "\ud835\udca2": "G",
+    "\u210B": "H",
+    "\u2110": "I",
+    "\ud835\udca5": "J",
+    "\ud835\udca6": "K",
+    "\u2112": "L",
+    "\u2113": "M",
+    "\ud835\udca9": "N",
+    "\ud835\udcaa": "O",
+    "\ud835\udcab": "P",
+    "\ud835\udcac": "Q",
+    "\u211B": "R",
+    "\ud835\udcae": "S",
+    "\ud835\udcaf": "T",
+    "\ud835\udcb0": "U",
+    "\ud835\udcb1": "V",
+    "\ud835\udcb2": "W",
+    "\ud835\udcb3": "X",
+    "\ud835\udcb4": "Y",
+    "\ud835\udcb5": "Z"
+  });
+
   // Mapping of Unicode accent characters to their LaTeX equivalent in text and
   // math mode (when they exist).
   var unicodeAccents = {
@@ -10766,6 +10790,22 @@ var temml = (function () {
             text
           };
         } else {
+          if (asciiFromScript[text]) {
+            // Unicode 14 disambiguates chancery from roundhand.
+            // See https://www.unicode.org/charts/PDF/U1D400.pdf
+            this.consume();
+            const nextCode = this.fetch().text.charCodeAt(0);
+            // mathcal is Temml default. Use mathscript if called for.
+            const font = nextCode === 0xfe01 ? "mathscr" : "mathcal";
+            if (nextCode === 0xfe00 || nextCode === 0xfe01) { this.consume(); }
+            return {
+              type: "font",
+              mode: "math",
+              font,
+              body: { type: "mathord", mode: "math", loc, text: asciiFromScript[text] }
+            }
+          }
+          // Default ord character. No disambiguation necessary.
           s = {
             type: group,
             mode: this.mode,
@@ -11022,7 +11062,7 @@ var temml = (function () {
    * https://mit-license.org/
    */
 
-  const version = "0.9.2";
+  const version = "0.10.0";
 
   function postProcess(block) {
     const labelMap = {};